Added functionality for classification.
TODO: implement bayesian classifier.

+++ distributionparameters.h
@@ -46,7 +46,7 @@ public:
         return b;
     }
 
-    inline double get_m()
+    inline int get_m()
     {
         return m;
     }
